requirements.txt
numpy
pandas
scikit-learn>=1.6.0
litellm
python-dotenv
mcp
fastmcp
tqdm
pyyaml
